Reorg a little
