+++ lib/gimme.js
@@ -1,4 +1,3 @@
-const Boom  = require('boom')
 const http  = require('http')
 const https = require('https')
 const qs    = require('qs')
@@ -10,7 +9,9 @@ const {
   length, merge, pick, pipe, replace, when
 } = require('ramda')
 
-const { parseBody, streamBody } = require('./body')
+const parseBody  = require('./parseBody')
+const streamBody = require('./streamBody')
+const wrapError  = require('./wrapError')
 
 const redact = evolve({ 'authorization': replace(/\s(.*)/, ' REDACTED') })
 
@@ -18,6 +19,8 @@ const clean = pipe(
   pick(['body', 'headers', 'statusCode', 'statusMessage']),
   evolve({ headers: redact })
 )
+
+const parseJSON = when(length, JSON.parse)
 
 const gimme = opts => {
   const { json = true } = opts
@@ -81,13 +84,6 @@ const gimme = opts => {
       req.end()
     }
   })
-}
-
-const parseJSON = when(length, JSON.parse)
-
-const wrapError = context => {
-  const { statusCode, statusMessage } = context.res
-  return Boom.create(statusCode, statusMessage, context)
 }
 
 module.exports = gimme
