+++ lib/parseBody.js
@@ -1,7 +1,7 @@
+const rawBody = require('raw-body')
+const typer   = require('media-typer')
+
 const { compose, curry, path } = require('ramda')
-const { PassThrough } = require('stream')
-const rawBody         = require('raw-body')
-const typer           = require('media-typer')
 
 const contentLength = compose(parseInt, path(['headers', 'content-length']))
 const contentType   = path(['headers', 'content-type'])
@@ -13,13 +13,4 @@ const parseBody = (deserialize, res) => {
   return rawBody(res, { encoding, length }).then(deserialize)
 }
 
-const streamBody = res => {
-  const body = new PassThrough()
-  res.pipe(body)
-  return Promise.resolve(body)
-}
-
-module.exports = {
-  parseBody: curry(parseBody),
-  streamBody
-}
+module.exports = curry(parseBody)
