+++ lib/wrapError.js
@@ -1,0 +1,8 @@
+const Boom = require('boom')
+
+const wrapError = context => {
+  const { statusCode, statusMessage } = context.res
+  return Boom.create(statusCode, statusMessage, context)
+}
+
+module.exports = wrapError
